requirements.txt
streamlit
genai
python-dotenv
